requirements.txt
attrs~=21.4.0